--- eunphyo/week2/src/main/java/com/example/week2/comment/entity/CommentReport.java
@@ -1,45 +1,0 @@
-package com.example.week2.comment.entity;
-
-import com.example.week2.global.entity.BaseEntity;
-import com.example.week2.report.entity.ReportStatus;
-import com.example.week2.user.entity.User;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "comment_reports")
-public class CommentReport extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id", nullable = false)
-    private Comment comment;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "reporter_id", nullable = false)
-    private User reporter;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String reason;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private ReportStatus status;
-
-    @Builder
-    public CommentReport(Comment comment, User reporter, String reason) {
-        this.comment = comment;
-        this.reporter = reporter;
-        this.reason = reason;
-        this.status = ReportStatus.PENDING;
-    }
-
-    public void resolve() {
-        this.status = ReportStatus.RESOLVED;
-    }
-}
