--- eunphyo/week2/src/main/java/com/example/week2/comment/entity/CommentReportStatus.java
@@ -1,7 +1,0 @@
-package com.example.week2.comment.entity;
-
-public enum CommentReportStatus {
-
-        PENDING,
-        RESOLVED
-}
