+++ eunphyo/week2/src/main/java/com/example/week2/comment/repository/CommentRepository.java
@@ -1,7 +1,16 @@
 package com.example.week2.comment.repository;
 
 import com.example.week2.comment.entity.Comment;
+import com.example.week2.comment.entity.CommentStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+import java.util.List;
+import java.util.Optional;
+
+public interface CommentRepository
+        extends JpaRepository<Comment, Long> {
+    Optional<Comment> findByIdAndStatus(
+            Long Id,
+            CommentStatus status
+    );
 }
