+++ eunphyo/week2/src/main/java/com/example/week2/comment/service/CommentService.java
@@ -3,6 +3,7 @@ package com.example.week2.comment.service;
 import com.example.week2.comment.dto.CommentCreateRequest;
 import com.example.week2.comment.dto.CommentResponse;
 import com.example.week2.comment.entity.Comment;
+import com.example.week2.comment.entity.CommentStatus;
 import com.example.week2.comment.repository.CommentRepository;
 import com.example.week2.post.entity.Post;
 import com.example.week2.post.repository.PostRepository;
@@ -13,6 +14,10 @@ import org.springframework.stereotype.Service;
 import com.example.week2.global.response.CustomException;
 import com.example.week2.global.response.ErrorCode;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+import static org.springframework.data.util.ClassUtils.ifPresent;
 
 @Service
 @RequiredArgsConstructor
@@ -59,18 +64,18 @@ public class CommentService {
 
     public CommentResponse.CommentDetailResponse getCommentResponse(Long commentId) {
 
-        Comment comment = commentRepository.findById(commentId)
+        Comment comment = commentRepository
+                .findByIdAndStatus(commentId, CommentStatus.ACTIVE)
                 .orElseThrow(() -> new CustomException(ErrorCode.COMMENT_NOT_FOUND));
 
         return CommentResponse.CommentDetailResponse.builder()
                 .commentId(comment.getId())
                 .content(comment.getContent())
-                .nickname(comment.getUser().getName()) // nickname이면 getNickname()
+                .nickname(comment.getUser().getName())
                 .createdAt(comment.getCreatedAt())
                 .updatedAt(comment.getUpdatedAt())
                 .build();
     }
-
 
 }
