+++ eunphyo/week2/src/main/java/com/example/week2/post/entity/Post.java
@@ -1,5 +1,6 @@
 package com.example.week2.post.entity;
 
+import com.example.week2.comment.entity.Comment;
 import com.example.week2.global.entity.BaseEntity;
 import com.example.week2.user.entity.User;
 import jakarta.persistence.*;
