+++ eunphyo/week2/src/main/java/com/example/week2/report/repository/ReportRepository.java
@@ -1,5 +1,6 @@
 package com.example.week2.report.repository;
 
+import com.example.week2.comment.entity.Comment;
 import com.example.week2.post.entity.Post;
 import com.example.week2.report.entity.Report;
 import com.example.week2.user.entity.User;
@@ -8,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface ReportRepository
         extends JpaRepository<Report, Long> {
 
-    boolean existsByPostAndUser(Post post, User user);
+    boolean existsByCommentAndReporter(Comment comment, User reporter);
 }
